Fix recipes to only use Conda-based tools
`file` tool is required to build `build gcc/newlib` and `gcc/linux-musl`
packages. It is only present in the `conda-forge` channel.

By adding `litex-hub` and `defaults` on top of the `condarc`'s channels,
also because of `channel_priority: strict` setting, only the packages
absent in these channels will be taken from the `conda-forge` channel.

The changes turned out to be enough to dump installing additional
packages through `brew` on macOS in all recipes but `sdcc` that failed
to built with tools installed through Conda.