lib/isl: Skip building due to missing sources tarball
The official tarball server has been down for a few days. It might be a
temporary problem but let's simply skip its building to keep the CI
green. The very same libisl-0.21 package is built each time with the
only differences coming from dependencies and Conda changes.

Perhaps the official servers will go back online until a new build of
this package is needed.

Related to: #34