gcc/newlib: Drop gcc-nostdc dependency
Only "new" files from PREFIX are packaged so after using gcc-nostdc as a
host package some of the crucial files were ignored. This was the root
cause behind the necessity to install gcc-newlib with gcc-nostdc.

The `${TOOLCHAIN_ARCH}-unknown-elf-gcc` and `${TOOLCHAIN_ARCH}-elf-gcc`
aren't found so these must've been previously installed with gcc-nostdc.